update types and rename components in UI
